Fix HTML injestions and pasted text parsing
