+++ src/web/utils/inputUtils.ts
@@ -31,4 +31,48 @@ function getElementHeight(node: HTMLDivElement, styles: CSSProperties, numberOfL
   return styles.height ? `${styles.height}px` : 'auto';
 }
 
-export {isEventComposing, getPlaceholderValue, getElementHeight};
+const parseInnerHTMLToText = (target: HTMLElement): string => {
+  let text = '';
+  const childNodes = target.childNodes ?? [];
+  childNodes.forEach((node, index) => {
+    const nodeCopy = node.cloneNode(true) as HTMLElement;
+    if (nodeCopy.innerHTML) {
+      // Replace single <br> created by contentEditable with '\n', to enable proper newline deletion on backspace, when next lines also have <br> tags
+      if (nodeCopy.innerHTML === '<br>') {
+        nodeCopy.innerHTML = '\n';
+      }
+      // Replace only br tags with data-id attribute, because we know that were created by the web parser. We need to ignore tags created by contentEditable div
+      nodeCopy.innerHTML = nodeCopy.innerHTML.replaceAll(/<br .*?>/g, '\n');
+    }
+
+    let nodeText = nodeCopy.textContent ?? '';
+    if (nodeCopy.innerHTML && nodeCopy.innerHTML.includes('<span data-type="paste"')) {
+      nodeText = nodeText.slice(0, -1);
+    }
+
+    // Remove unnecessary new lines from the end of the text
+    if (nodeText.length > 2 && nodeText[-3] !== '\n' && nodeText.slice(-2) === '\n\n') {
+      nodeText = nodeText.slice(0, -1);
+    }
+
+    // Last line specific handling
+    if (index === childNodes.length - 1) {
+      if (nodeText === '\n\n') {
+        // New line creation
+        nodeText = '\n';
+      } else if (nodeText === '\n') {
+        // New line deletion on backspace
+        nodeText = '';
+      }
+    }
+
+    text += nodeText;
+    // Split paragraphs with new lines
+    if (/[^\n]/.test(nodeText) && index < childNodes.length - 1) {
+      text += '\n';
+    }
+  });
+  return text;
+};
+
+export {isEventComposing, getPlaceholderValue, getElementHeight, parseInnerHTMLToText};
